+++ sdk/appcenter-distribute/src/test/java/com/microsoft/appcenter/distribute/download/manager/DownloadManagerReleaseDownloaderTest.java
@@ -301,13 +301,12 @@ public class DownloadManagerReleaseDownloaderTest {
     public void completeDownload() {
         mockStatic(Uri.class);
         when(Uri.parse(anyString())).thenReturn(mock(Uri.class));
-        when(mListener.onComplete(any(Uri.class))).thenReturn(true);
 
         /* Complete download. */
-        mReleaseDownloader.onDownloadComplete(mock(Cursor.class));
+        mReleaseDownloader.onDownloadComplete();
 
         /* Verify. */
-        verify(mListener).onComplete(any(Uri.class));
+        verify(mListener).onComplete(anyLong());
         verify(mListener, never()).onError(anyString());
     }
 
@@ -315,30 +314,14 @@ public class DownloadManagerReleaseDownloaderTest {
     public void completeDownloadFallbackOnOldDevices() {
         mockStatic(Uri.class);
         when(Uri.parse(anyString())).thenReturn(mock(Uri.class));
-        when(mListener.onComplete(any(Uri.class))).thenReturn(false).thenReturn(true);
         Whitebox.setInternalState(Build.VERSION.class, "SDK_INT", Build.VERSION_CODES.M);
 
         /* Complete download. */
-        mReleaseDownloader.onDownloadComplete(mock(Cursor.class));
+        mReleaseDownloader.onDownloadComplete();
 
         /* Verify. */
-        verify(mListener, times(2)).onComplete(any(Uri.class));
+        verify(mListener).onComplete(anyLong());
         verify(mListener, never()).onError(anyString());
-    }
-
-    @Test
-    public void completeDownloadingFallbackOnNewDevices() {
-        mockStatic(Uri.class);
-        when(Uri.parse(anyString())).thenReturn(mock(Uri.class));
-        when(mListener.onComplete(any(Uri.class))).thenReturn(false).thenReturn(true);
-        Whitebox.setInternalState(Build.VERSION.class, "SDK_INT", Build.VERSION_CODES.O);
-
-        /* Complete download. */
-        mReleaseDownloader.onDownloadComplete(mock(Cursor.class));
-
-        /* Verify. */
-        verify(mListener).onComplete(any(Uri.class));
-        verify(mListener).onError(anyString());
     }
 
     @Test
@@ -348,10 +331,10 @@ public class DownloadManagerReleaseDownloaderTest {
 
         /* Complete download after cancelling. */
         mReleaseDownloader.cancel();
-        mReleaseDownloader.onDownloadComplete(mock(Cursor.class));
+        mReleaseDownloader.onDownloadComplete();
 
         /* Verify. */
-        verify(mListener, never()).onComplete(any(Uri.class));
+        verify(mListener, never()).onComplete(anyLong());
     }
 
     @Test